update button props
