+++ packages/button/src/BaseButton.tsx
@@ -4,10 +4,30 @@ import {
 	forwardRef,
 	useCallback,
 	useRef,
+	PropsWithChildren,
 } from 'react';
 import { mergeRefs, tokens } from '@ag.ds-next/core';
 
-export type BaseButtonProps = ButtonHTMLAttributes<HTMLButtonElement>;
+type NativeButtonProps = ButtonHTMLAttributes<HTMLButtonElement>;
+
+export type BaseButtonProps = PropsWithChildren<{
+	/** Identifies the element (or elements) whose contents or presence are controlled by the current element. */
+	'aria-controls'?: NativeButtonProps['aria-controls'];
+	/** Identifies the element (or elements) that describes the object. */
+	'aria-describedby'?: NativeButtonProps['aria-describedby'];
+	/** Indicates whether the element, or another grouping element it controls, is currently expanded or collapsed. */
+	'aria-expanded'?: NativeButtonProps['aria-expanded'];
+	/** Describes the button element to assistive technologies. */
+	'aria-label'?: NativeButtonProps['aria-label'];
+	/** The CSS class name, typically generated from the `css` prop. */
+	className?: string;
+	/** When true, prevents onClick from firing. */
+	disabled?: boolean;
+	/** Function to be fired following a click event of the button. */
+	onClick?: NativeButtonProps['onClick'];
+	/** Provide an alternate type if the button is within a form. */
+	type?: NativeButtonProps['type'];
+}>;
 
 export const BaseButton = forwardRef<HTMLButtonElement, BaseButtonProps>(
 	function BaseButton({ onClick: onClickProp, ...props }, forwardedRef) {
